Conflicts:
	src/qml/jsruntime/qv4arraydata.cpp
	src/qml/jsruntime/qv4context_p.h
	src/qml/jsruntime/qv4globalobject.cpp
	src/qml/jsruntime/qv4internalclass.cpp
	src/quick/items/qquicktext_p.h
	src/quick/items/qquicktextedit_p.h
	src/quick/items/qquicktextinput_p.h

+++ src/quick/items/qquickwindow.h
@@ -143,7 +143,7 @@ public:
 
     void scheduleRenderJob(QRunnable *job, RenderStage schedule);
 
-    int effectiveDevicePixelRatio() const;
+    qreal effectiveDevicePixelRatio() const;
 
 Q_SIGNALS:
     void frameSwapped();
